requirements.txt
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
transformers==4.38.2
sentence-transformers==2.6.1
faiss-cpu
fastapi
uvicorn
